requirements.txt
accelerate==1.5.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
easydict==1.13
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-babel==4.0.0
flask-cors==5.0.1
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gast==0.3.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
paddleclas==2.6.0
paddleocr==2.10.0
paddlepaddle==0.0.0
pandas==2.2.3
pillow==11.1.0
premailer==3.10.0
prettytable==3.15.1
protobuf==6.30.1
psutil==7.0.0
pyclipper==1.3.0.post6
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.3.13
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@0ebd6651acd32c982fee265b23243b89bdb89577
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.3
